requirements.txt
piexif==1.1.1
pillow=10.0.1